backend/agents/workItemAgent/pyproject.toml
[project]
name = "work-item-agent"
version = "1.0.0"
description = "Azure DevOps Work Item Agent with OTEL observability"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.25.0",
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "aws-opentelemetry-distro>=0.15.0",
    "bedrock-agentcore>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
